# 09_K-API-09_API_REFERENCE/openapi/kubric-noc-api.yaml
openapi: "3.1.0"
info:
  title: "Kubric NOC API"
  version: "1.0.0"
  description: |
    Network Operations Center API for the Kubric Unified Detection & Response platform.

    ## Authentication
    All endpoints require a Bearer JWT issued by the Kubric Identity service.
    The JWT must contain the `tenant_id` claim and the appropriate RBAC role
    (`noc:read`, `noc:write`, or `noc:admin`).

    ## Tenant isolation
    Every request is scoped to the tenant encoded in the JWT. Cross-tenant
    access is only possible with the `admin` super-role.
  contact:
    name: "Kubric Engineering"
    url: "https://kubric.io"
  license:
    name: "Proprietary"
servers:
  - url: "https://api.kubric.io/noc"
    description: "Production"
  - url: "http://localhost:8083"
    description: "Development"
...
security:
  - BearerAuth: []
paths:
...
  /detection/incidents:
    get:
      operationId: listIncidents
      tags: [detection]
      summary: List detection incidents
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: severity
          in: query
          schema:
            type: string
            enum: [critical, high, medium, low, info]
        - name: status
          in: query
          schema:
            type: string
            enum: [open, investigating, contained, resolved, closed, false_positive]
        - name: from
          in: query
          schema:
            type: string
            format: date-time
          description: Return incidents created at or after this timestamp (RFC 3339)
        - name: to
          in: query
          schema:
            type: string
            format: date-time
          description: Return incidents created before this timestamp (RFC 3339)
      responses:
        "200":
          description: Paginated incident list
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Incident"
                  meta:
                    $ref: "#/components/schemas/PaginationMeta"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /detection/incidents/{id}:
    parameters:
      - $ref: "#/components/parameters/IDParam"
    get:
      operationId: getIncident
      tags: [detection]
      summary: Get full incident details
      responses:
        "200":
          description: Incident with full detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateIncident
      tags: [detection]
      summary: Update incident status, assignee, or notes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidentUpdateRequest"
      responses:
        "200":
          description: Updated incident
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Incident"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /detection/incidents/{id}/dispatch:
    parameters:
      - $ref: "#/components/parameters/IDParam"
    post:
      operationId: dispatchIncident
      tags: [detection]
      summary: Dispatch incident to an integration (TheHive, PagerDuty, etc.)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target]
              properties:
                target:
                  type: string
                  enum: [thehive, pagerduty, jira, slack, email]
                  description: Integration target
                priority:
                  type: string
                  enum: [P1, P2, P3, P4]
                  default: P2
                message:
                  type: string
                  description: Optional supplementary note
      responses:
        "202":
          description: Dispatch accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  dispatch_id:
                    type: string
                    format: uuid
                  target:
                    type: string
                  external_ref:
                    type: string
                    description: Reference ID in the target system
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /detection/timeline:
    get:
      operationId: getTimeline
      tags: [detection]
      summary: Retrieve the detection event timeline
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: agent_id
          in: query
          schema:
            type: string
            format: uuid
        - name: incident_id
          in: query
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Ordered timeline events
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items:
                      $ref: "#/components/schemas/TimelineEvent"
        "400":
          $ref: "#/components/responses/BadRequest"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IDParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Resource UUID
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 50
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequest:
      description: Invalid request payload
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable description
        details:
          type: object
          additionalProperties: true
    PaginationMeta:
      type: object
      required: [page, limit, total]
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        has_more:
          type: boolean
...
    Incident:
      type: object
      required: [id, tenant_id, title, severity, status, source, detected_at]
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [critical, high, medium, low, info]
        status:
          type: string
          enum: [open, investigating, contained, resolved, closed, false_positive]
        source:
          type: string
          description: Detection source (e.g. sigma, yara, network_ids)
        mitre_tactics:
          type: array
          items:
            type: string
          description: MITRE ATT&CK tactic IDs (e.g. TA0001)
        mitre_techniques:
          type: array
          items:
            type: string
          description: MITRE ATT&CK technique IDs (e.g. T1059.001)
        affected_agents:
          type: array
          items:
            type: string
            format: uuid
        assignee:
          type: string
        tags:
          type: array
          items:
            type: string
        external_refs:
          type: object
          additionalProperties:
            type: string
          description: Cross-references to external systems (e.g. thehive_case_id)
        detected_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        resolved_at:
          type: string
          format: date-time
    IncidentUpdateRequest:
      type: object
      properties:
        status:
          type: string
          enum: [investigating, contained, resolved, closed, false_positive]
        assignee:
          type: string
        tags:
          type: array
          items:
            type: string
        notes:
          type: string
    TimelineEvent:
      type: object
      required: [id, event_type, ocsf_class, agent_id, timestamp]
      properties:
        id:
          type: string
          format: uuid
        event_type:
          type: string
          description: Kubric event category (process, file, network, auth, detection)
        ocsf_class:
          type: integer
          description: OCSF class ID (e.g. 4007 = Process Activity)
        agent_id:
          type: string
          format: uuid
        incident_id:
          type: string
          format: uuid
          description: Correlated incident if applicable
        severity:
          type: string
          enum: [critical, high, medium, low, info]
        summary:
          type: string
        raw_data:
          type: object
          description: Original event payload (OCSF-normalised)
        timestamp:
          type: string
          format: date-time
...